refactor: Wrap Identity in its own struct
This wrapper includes platform and status in the struct belonging to the identity if ever fetched from a multipass extension.

Note: This dereferences back to warp `Identity`, which can allow access to the getters (and setters, if needed) while also giving access to the enum for the online status and platform in the wrapper. . Identity from warp can be converted into this by using `Identity::from`.