Relicense to MIT
The current CCC license is not appropriate for software and it would be
best to have a proper license before any external parties contribute.
Otherwise, we would have to ask their permission before re-licensing.

After discussion with the team, we have decided to relicense the project to MIT.

I just copied the `LICENSE` file from the `Warp` repo. I think we want
it to be kept in sync with the `Warp` repo.